pyproject.toml
[tool.poetry]
name = "pyro"
version = "0.1.0"
description = "PYthon Reader & OCR"
authors = ["Pingu <ldaniel358@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
gunicorn = "^22.0.0"
python-dotenv = "^1.0.1"
flask-cors = "^4.0.1"
flask-restful = "^0.3.10"
boto3 = "^1.34.103"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
pylint = "^3.1.0"
mypy = "^1.10.0"
pytest = "^8.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
